requirements.txt
Flask==3.1.2
pypdf==6.6.0
Pillow==12.1.0
reportlab>=4.0.0
